Refactor Character Distributor UI and enhance modular structure
- Update entry point and import structure for better organization
- Introduce modular code structure following the Single Responsibility Principle
- Improve authentication flow with enhanced token management and error handling
- Add event handler setup for improved UI interactions
- Update README to reflect new code structure and organization
- Include ESLint as a development dependency for code quality assurance
